+++ packages/supacloud-js/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.18.3](https://github.com/zuohuadong/supacloud/compare/supacloud-js-v0.18.2...supacloud-js-v0.18.3) (2026-06-03)
+
+
+### Miscellaneous Chores
+
+* upgrade all dependencies to latest minor ([e9719d9](https://github.com/zuohuadong/supacloud/commit/e9719d983c9303c783ddcd3d772c7e7c56e985b9))
+
 ## [0.18.2](https://github.com/zuohuadong/supacloud/compare/supacloud-js-v0.18.1...supacloud-js-v0.18.2) (2026-05-29)